pyproject.toml
[project]
name = "unicef-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "google-auth>=2.48.0",
    "hl7>=0.4.5",
    "hl7apy>=1.3.5",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
